redis_vizjson_cache moved from app/helpers to lib/cartodb
