+++ app/controllers/carto/api/vizjson3_presenter.rb
@@ -1,4 +1,5 @@
 require_dependency 'carto/api/layer_vizjson_adapter'
+require_dependency 'cartodb/redis_vizjson_cache'
 
 module Carto
   module Api
