+++ app/controllers/carto/api/vizjson_presenter.rb
@@ -1,6 +1,6 @@
 require_dependency 'visualization/vizjson'
 require_dependency 'carto/api/visualization_vizjson_adapter'
-require_relative '../../../helpers/redis_vizjson_cache'
+require_dependency 'cartodb/redis_vizjson_cache'
 
 module Carto
   module Api
