+++ lib/cartodb/redis_vizjson_cache.rb
